Add Mail Service

+++ src/routes/admin.js
@@ -1,19 +1,19 @@
 const express = require('express');
 const Medical = require("../models/medical");
+const Doctor = require('../models/doctor');
 const Lab = require("../models/lab");
-const Doctor = require("../models/doctor");
+const { mailadminappoved } = require('../utils/mailservice');
 const router = new express.Router();
 
 
 //Fetching Pending Doctors, Medicals ans Labs
 router.get('/pendingdoctors',async (req,res)=>{
     try {
-        const PendingDoctors = await Doctor.find({approved:false});
-        console.log(PendingDoctors);    
+        const PendingDoctors = await Doctor.find({approved:false})
         res.status(200).send(PendingDoctors);
         
-    } catch (e) {
-        req.statu(500).send(e.message)
+    } catch (err) {
+        res.status(500).send(err.message)
     }
 })
 
@@ -23,8 +23,8 @@ router.get('/pendinglabs',async (req,res)=>{
         console.log(PendingLabs);    
         res.status(200).send(PendingLabs);
         
-    } catch (e) {
-        req.statu(500).send(e.message)
+    }catch (err) {
+        res.status(500).send(err.message)
     }
 })
 
@@ -35,7 +35,7 @@ router.get('/pendingMedicals',async (req,res)=>{
         res.status(200).send(PendingMedicals);
         
     } catch (e) {
-        req.statu(500).send(e.message)
+        res.status(500).send(e.message)
     }
 })
 
@@ -46,11 +46,19 @@ router.patch('/doctorapproved',async (req,res) => {
 
     const {doctorid,approved} = req.body; //approved --> 1,rejection --> 0
     try {
+        
+        const doctor = await Doctor.findOne({doctorid,approved:false});
+        if(!doctor){
+            throw new Error("Doctor Not Found")
+        }
+
         if(approved){
             await Doctor.updateOne({doctorid},{$set:{approved : true}});
+            mailadminappoved(doctor.contacts.email,doctor.details.fname,doctor.details.lname,doctorid,1);
             res.status(200).send("Successfully Approved");
         }else{
             await Doctor.deleteOne({doctorid});
+            mailadminappoved(doctor.contacts.email,doctor.details.fname,doctor.details.lname,doctorid,0);
             res.status(200).send("Successfully DisApproved");
         }
     } catch (e) {
@@ -61,13 +69,19 @@ router.patch('/doctorapproved',async (req,res) => {
 
 router.patch('/labapproved',async (req,res) => {
     const {labid,approved} = req.body; //approved --> 1,rejection --> 0
+    const lab = await Lab.findOne({labid,approved:false})
+    if(!lab){
+        throw new Error('Lab Not Found');
+    }
     try {
         if(approved){
             await Lab.updateOne({labid},{$set:{approved : true}});
+            mailadminappoved(lab.contacts.email,lab.ownerdetails.fname,lab.ownerdetails.lname,labid,1);
             res.status(200).send("Successfully Approved");
         }else{
             await Lab.deleteOne({labid});
             res.status(200).send("Successfully DisApproved");
+            mailadminappoved(lab.contacts.email,lab.ownerdetails.fname,lab.ownerdetails.lname,labid,0);
         }
     } catch (e) {
         res.status(500).send(e.message);
@@ -77,11 +91,18 @@ router.patch('/labapproved',async (req,res) => {
 router.patch('/medicalapproved',async (req,res) => {
     const {medicalid,approved} = req.body; //approved --> 1,rejection --> 0
     try {
+        const medical = await Medical.findOne({medicalid,approved : false});
+        if(!medical){
+            throw new Error('Medical is Not Found');
+        }
+
         if(approved){
             await Medical.updateOne({medicalid},{$set:{approved : true}});
+            mailadminappoved(medical.contacts.email,medical.ownerdetails.fname,medical.ownerdetails.lname,medicalid,1);
             res.status(200).send("Successfully Approved");
         }else{
             await Medical.deleteOne({medicalid});
+            mailadminappoved(medical.contacts.email,medical.ownerdetails.fname,medical.ownerdetails.lname,medicalid,0);
             res.status(200).send("Successfully DisApproved");
         }
     } catch (e) {
